backend/pyproject.toml
[tool.poetry]
name = "video-frame-extractor"
version = "0.1.0"
description = "Video frame extraction service"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"
ffmpeg-python = "^0.2.0"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.10.0"
isort = "^5.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 